airbnb/requirements.txt
playwright==1.47.0
jinja2==3.1.4
PyYAML==6.0.2
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.2.2
